Add .gitignore file
This commit introduces a .gitignore file to the project. This file is used to specify which files and directories to ignore, so certain files (like temporary files, logs, etc.) aren't tracked by Git. In this case, it covers a wide range of cases specific to Python development, including patterns for common Python tools and libraries, as well as various IDE and tool-specific files.