Add Zappa support for deployment
Added 'zappa' to the list of requirements in the requirements.txt file and created a zappa_settings.json configuration file for easier deployment of the application to a serverless AWS environment. The configuration is set for a "production" environment with a python3.10 runtime and a specific S3 bucket for storing serverless application files.